Store everything in the DB, query it to print information
This change makes it much faster to repeatedly query the CircleCI API because it will only fetch data when the API indicates that it has changed. It now always uses the database as a source of information, which has the advantage of letting us search through historical data using the command line instead of needing to use the API for most activity.