Add task usage stats
The main user-facing feature is downloading task usage stats from the endpoint that CircleCI uses for its own charts in the web UI. Currently only the maximum CPU and RAM usage across all jobs in a workflow are stored in a JSON doc next to the binary for further processing. Useful for finding jobs that hit 100% RAM usage, which could point to where OOMs are causing test failures or very low CPU and RAM usage, indicating that the executor being used is larger than it needs to be.